% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicsaudit-package.R
\docType{package}
\name{omicsaudit-package}
\alias{omicsaudit}
\alias{omicsaudit-package}
\title{omicsaudit: Black-Box Audit Toolkit for Omics Biomarker Pipelines}
\description{
Audits biomarker analysis pipelines by controlling the data that goes in, using planted signals. Also provides a zoo of deliberately flawed pipelines with known answers.
}
\author{
\strong{Maintainer}: omicsaudit authors \email{benabanks@icloud.com}

}
\keyword{internal}
