% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo.R
\name{zoo_dataset}
\alias{zoo_dataset}
\title{Generate the dataset a zoo entry is defined on}
\usage{
zoo_dataset(name)
}
\arguments{
\item{name}{Entry name, or a list returned by \code{\link[=zoo_load]{zoo_load()}}.}
}
\value{
An \code{omics_sim} object (see \code{\link[=simulate_omics]{simulate_omics()}}).
}
\description{
Generate the dataset a zoo entry is defined on
}
